rename : renamed classes related to AcceptanceTest
